Create directory for mixins
